Factor out createStreetList().
